refactor particles to grid into particles module
