+++ setup.py
@@ -33,7 +33,7 @@ setup(name='postpic',
       description='The open source particle-in-cell post processor.',
       url='https://github.com/skuschel/postpic',
       packages=find_packages(include=['postpic*']),
-      ext_modules = cythonize("postpic/cythonfunctions.pyx"),
+      ext_modules = cythonize("postpic/particles/_particlestogrid.pyx"),
       include_dirs = [numpy.get_include()],
       license='GPLv3+',
       setup_requires=['cython>=0.18'],
